+++ Controller/Produto.cs
@@ -10,7 +10,7 @@ namespace GM.Controller {
 		public int familia {get; set;}
 		public String unidade {get; set;}
 		public String cod_barra {get; set;}
-		public decimal preco {get; set;}
+		public float preco {get; set;}
 		public int situacao {get; set;} // 0 - Suspenso / 1 - Ativo
 		
 	}
